Fix some bugs in kernel isp code.
In isp_user.h, the struct for the isphist region parameters was the
wrong size given the way it is used in isphist.c. Params need to be
32 bit.

The isphist stats were off by one because of a write to the data reg
after setting params for no reason. Obviously, code that was never
tested.

Cleaned up some code in the isp hist set params function to correctly
or at least simplify the validation of num bins vs num regions.